Give the same weavedns IP to all containers (bf416b4df836d76106d11c6f13113fd93f8dd04a)
Since we're routing through the weave bridge, it doesn't matter which
IP we supply as `--dns`. It's sometimes easier to supply the same
address e.g., via swarm, where we don't know where something is
running. All this takes is to give each container a route to the right
place (in fact, it might be even better to give a subnet ...)